backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
drf-yasg==1.21.7
django-cors-headers==4.3.1 